Use var envs for the Geolite DB url and update interval
We've update the URL that gets the database containing the geolocation
data.
We need this because the Freegeoip relies on the GeoLite2 databases to get the
location data and since they are updating the way they serve it we need
to comply with the new updates that went live on the 20191230. More info
check the Maxmind blog entry
https://blog.maxmind.com/2019/12/18/significant-changes-to-accessing-and-using-geolite2-databases/